Add tocfl.xlsx
Add XLSX file containing vocabulary from Taiwan's TOCFL (Test of Chinese
as a Foreign Language).

This file was downloaded from
https://www.sc-top.org.tw/english/download.php. It is the "8000 Chinese
Words" link. This downloads a zip file, and the XLSX file is inside.
Note that the folders inside the zip file have filenames that aren't
valid Unicode (probably Big5), but you can still extract the file and
rename it to a valid name.